requirements_web.txt
# Core dependencies
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
